requirements.txt
absl-py==0.11.0
asgiref==3.2.10
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
Django==3.0.5
django-cors-headers==3.5.0
django-rest-knox==4.1.0
djangorestframework==3.12.1
djongo==1.3.3
dnspython==2.1.0
gast
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio~=1.34.0
h5py
idna==2.10
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.3.3
numpy
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.11.2
pytz==2020.1
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy
six==1.15.0
smmap==3.0.4
sqlparse==0.2.4
tensorboard
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
